re-organize a bit the code
- move TelemetryConfig struct along with the PluginConfigSchemaV1 struct config
- move telemetry provider graphite implementation to its own file as well as
the corresponding tests
